Move EACRugged IOControl code to specific package
